fix(mobile): rename users API to .ts and use auth barrel import
- Rename services/api/users.tsx -> users.ts (no JSX in the file).
- Import loadAuth from '../../auth' via the barrel; extend the barrel
  to re-export loadAuth, saveAuth, and the AuthResult type so other
  call sites can follow the same convention.
